cmd/go/internal/web: split interceptor into separate package
This moves the interception code ito package
cmd/go/internal/web/intercept so that it can also be used by
cmd/go/internal/auth.

For #26232

Reviewed-on: https://go-review.googlesource.com/c/go/+/625036
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
